Crystal Ability & Some Adjustments
Adjustments:

Fixed jump system not working properly when the Player jumps while moving.
Double Jump state has been created and made a future.
Player's invisibility in Dash ability has been made a future.
A system has been added to Clone ability to create a copy of the cole at a certain probability and distance.

Crystal Ability:

Player when press F key, crystals that go towards the enemy and cause explosion damage have been added.
Crystals are combined with other abilities.
Crystals appear instead of clones according to the player's preference.
If moving towards the enemy is not selected in the Crystal Ability, Player drops a crystal where it is and when presses F key again, it teleports to the crystal and creates a clone in the other position.
In the Blackhole Ability, it sends crystals instead of creating clones and unlike the clone, Player does not disappear while sending crystals.