Items & Inventory Section
1-> The inventory system has been set up for preparation. It will be redesigned with UI design in the future.
2-> A stash system has been set up to allow some items to be stacked.
3-> Items are divided into two categories: materials and equipment. Craftable materials will be added in the future.
4-> Enemies drop items with randomly assigned values when they die.
5-> The character drops items from the inventory based on their random values.
6-> To prevent these abilities from conflicting with each other, the character cannot equip more than one armor, sword, or jewelry item at the same time.
7-> For now, there are potions that provide armor, health, and fire damage. These potions are currently activated with the "Left Alt" key.
8-> The attributes of items classified as Epic Items vary. The rarity of these items is significantly higher than others. The currently added epic items and their features are:

Amulet Of God = When equipped with this item, our crystal and sword-throwing skills additionally perform a special lightning attack and shock ability.
Blue Wolf Armor = When equipped with this item, the character freezes nearby enemies for 5 seconds when their health reaches approximately 1 to 5 levels.
Rogue's Jacket = When equipped with this item, the character's magic resistance, health, lightning damage, and evasion attributes are enhanced.
Dracula's Sword = When equipped with this sword, the character regenerates a small amount of health with each hit on enemies.
IceLava Queen's Sword = When equipped with this sword, the character slows down the enemy with ice damage on each hit and activates the power of ice and flame on the third combo attack, dealing explosive ice-lava mixture damage to enemies.
Thunder Claw = When equipped with this sword, the character deals shock damage to enemies with each hit and performs a special lightning attack and shock ability with every strike.